Fix some bugs in delete and get stats
